SQUASH:
Install project in bootstrap commands and place project in pkg_self/

Replace platform arch check for batch with environment flag

Remove @install_project

Replace @pip's is_task_local with correct version

Remove bad util funs around local task checking

Refactor and mark AWS tests to not fail locally (and pass linter)

Add shebangs to shell scripts

TEMP narrow tests to speed up CI debugging

Revert "TEMP narrow tests to speed up CI debugging"

This reverts commit 5cae4e98eebe08dd34e9e2904542c0cd05728873.

Don't add project env bootstrap commands if default env is local

changed to pip install -e

add alternative for production

include mysqldb config

Big refactor...

General

- Simply GH actions - use AWS secrets manager; remove uneeded installs
- DOCUMENTATION - README.md provides a reasonable starting point by use
- Do not modify default --package-suffixes - let users decide and avoid
  potential security issues by unexpectedly sharing sensitive config
- Pin to metaflow<2.5 due to breaking changes in the extensions API
- Tidy-up tests; better matching them to behaviours under test; and
  changing them based on changed behaviour

@pip

- Added `safe` boolean toggle - provides speed in exchange for safety
  during development.

Project environment

- `project` environment is no longer set to default
- No longer install a local package automatically - @pip should be used
  instead if needed
- Export $MFPYTHON bash variable for using the right Python executable
  in preinstall scripts
- Preinstall packages feature removed - this can be done using `
  $MFPYTHON` in a `preinstall-*.sh` script (will not work locally on
  Conda! - Too complex and error-prone)
- Job package is no longer added to `pkg_self/` - this means we don't
  need to worry about `PYTHONPATH` or `@pip` having inconsistent
  behaviour locally vs remote BUT we do have to run locally from the
  project-level directory and be careful that there are no filename
  collisions.

Replace daps-utils with nuts-finder to reduce assumptions needed

Replace nuts-finder with tqdm because it doesn't require system deps

Make changes to be compatible with metaflow>=2.5.2

Run CI tests on all PRs

SQUASH...

Install project in bootstrap commands and place project in pkg_self/

Replace platform arch check for batch with environment flag

Remove @install_project

Replace @pip's is_task_local with correct version

Remove bad util funs around local task checking

Refactor and mark AWS tests to not fail locally (and pass linter)

Add shebangs to shell scripts

TEMP narrow tests to speed up CI debugging

Revert "TEMP narrow tests to speed up CI debugging"

This reverts commit 5cae4e98eebe08dd34e9e2904542c0cd05728873.

Don't add project env bootstrap commands if default env is local

changed to pip install -e

add alternative for production

include mysqldb config

Encapsulate bootstrap fallback logic

Add env var for python bin to be used by Metaflow

Allow mflog to be used in subshells

Strip down functionality to avoid complexity

Add unsafe option for @pip+conda

Add /metaflow/pkg_self to PYTHONPATH in bootstrap_commands

Have to warn rather than error on pip check because PYTHONPATH is checked

Do not change default environment - users should opt-in to complexity

WIP

- README draft
- Get rid of pkg_self/ in job package  - too hard to resolve paths in different envs
- Remove package suffix defaults
- Fix safe mode attr in @pip
- Leaner cmds in project bootstrap

WIP

WIP

Use AWS secrets manager for metaflow config

Use metaflow_extensions patch metaflow@ext_packaging

https://github.com/Netflix/metaflow/pull/959

Remove unneeded GHA steps

Version number

WIP: align with 25_fix_project_env changes for metaflow<2.5

Version modifier

Remove metaflow from requirements.txt

Update comments on extensions mechanism

Don't acutally use dap-utils mixin because it makes extra assumptions
that break test isolation

Replace daps-utils with nuts-finder to reduce assumptions needed

Dont run CI on macos until https://github.com/Netflix/metaflow/issues/969 sorted

Replace nuts-finder with tqdm because it doesn't require system deps

Refactor is_mflow_conda_environment to avoid mock

Fix requirements.txt for project packaging

Trying to get extensions working properly

Workaround macos + conda + batch issue

Try 2.5.2

Revert "Try 2.5.2"

This reverts commit bef33d5680f7461d7e44d13915a1dd7da08190d1.

minor doc tweak

Trigger CI on all PRs

Try latest release branch

Revert "Try latest release branch"

This reverts commit 923c80af4e042519a2dcc4f7b7ddd3dc994bf700.

Try latest release branch

Update secrets name
