+++ .github/workflows/tests.yml
@@ -56,4 +56,4 @@ jobs:
 
       - name: Run pytest
         run: |
-          $CONDA/bin/python -m pytest -m ""
+          export CONDA_SAFETY_CHECKS=disabled && $CONDA/bin/python -m pytest -m ""
