+++ setup.py
@@ -9,7 +9,7 @@ setup(
     install_requires=REQUIRES,
     python_requires=">=3.7",
     packages=find_namespace_packages(include=["metaflow_extensions.*"]),
-    version="0.2.1",
+    version="0.3.0",
     description="Nesta plugins for Metaflow (metaflow.org)",
     author="Nesta",
     license="MIT",
