+++ tests/conftest.py
@@ -5,8 +5,8 @@ from pathlib import Path
 
 import pytest
 
-
 from metaflow_extensions.nesta.utils import pip_install
+
 from utils import remove_pkg  # noqa: I
 
 MYPROJECT_PATH = Path(__file__).parent / "myproject"
@@ -16,7 +16,7 @@ MYPROJECT_PATH = Path(__file__).parent / "myproject"
 def temporary_project_maker(tmpdir_factory, project_name):
     project_path = Path(__file__).parent / project_name
     temp = tmpdir_factory.mktemp("data-project")
-    shutil.copytree(project_path, temp / project_name)
+    shutil.copytree(project_path, temp / project_name, symlinks=True)
     pip_install(sys.executable, "tqdm")
     yield temp
     remove_pkg("tqdm")
@@ -27,13 +27,6 @@ def temporary_project_maker(tmpdir_factory, project_name):
 def temporary_project(tmpdir_factory):
     """Create temporary project `myproject`."""
     with temporary_project_maker(tmpdir_factory, "myproject") as temp:
-        yield temp
-
-
-@pytest.fixture(scope="session")
-def temporary_duffproject(tmpdir_factory):
-    """Create temporary project `myduffproject`."""
-    with temporary_project_maker(tmpdir_factory, "myduffproject") as temp:
         yield temp
