+++ tests/myproject/myproject/flows/myproject
@@ -1,0 +1,1 @@
+../../../myproject
