+++ tests/myproject/myproject/flows/requirements-project_packaging_flow.txt
@@ -1,1 +1,1 @@
-.
+-e myproject/
